requirements.txt
aiohttp==1.0.5
async-timeout==3.0.0
chardet==3.0.4
discord==0.0.2
discord.py==0.16.12
multidict==4.3.1
PyYAML==3.12
websockets==3.4
